auth/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pytesseract
pdf2image
Pillow
